backend/django_back/requirements_api.txt
# Django et REST Framework
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7

# Base de données et ORM
# (SQLite inclus avec Python)

# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=4.9.0

# Analyse de texte
python-dotenv>=1.0.0

# Utilitaires
python-dateutil>=2.8.2
